site/saints/requirements.txt
django-cors-headers>=4.3,<5.0
# Additional dependencies
django-cron>=0.6
django-debug-toolbar>=4.4,<5.0
# ...existing code...
django-nested-admin>=4.0
djangorestframework>=3.14,<4.0
drf-spectacular>=0.27,<1.0
feedgen>=0.11
# ...existing code...
